inval case and expect more case proof

+++ src/strtoimax/IntNotations.v
@@ -19,18 +19,18 @@ Infix "%" := Int.mods (at level 70) : IntScope.
 Infix "//" := Int.divs (at level 70) : IntScope.
 
 Delimit Scope Int64Scope with int64.
-Infix "==" := Int.eq (at level 70) : Int64Scope.
-Notation "x ~= y" := (negb Int.eq x y) (at level 70) : Int64Scope.
-Notation "x >> y" := (Int.shru x y) (at level 70) : Int64Scope.
-Notation "0" := Int.zero : Int64Scope.
-Notation "1" := Int.one : Int64Scope.
-Infix "+" := Int.add : Int64Scope.
-Infix "-" := Int.sub : Int64Scope.
-Infix "*" := Int.mul : Int64Scope.
-Infix "<" := Int.lt : Int64Scope.
-Notation "x <= y" := (negb (Int.ltu y x)) (at level 70) : Int64Scope.
-Infix "%" := Int.mods (at level 70) : Int64Scope.
-Infix "//" := Int.divs (at level 70) : Int64Scope.
+Infix "==" := Int64.eq (at level 70) : Int64Scope.
+Notation "x ~= y" := (negb Int64.eq x y) (at level 70) : Int64Scope.
+Notation "x >> y" := (Int64.shru x y) (at level 70) : Int64Scope.
+Notation "0" := Int64.zero : Int64Scope.
+Notation "1" := Int64.one : Int64Scope.
+Infix "+" := Int64.add : Int64Scope.
+Infix "-" := Int64.sub : Int64Scope.
+Infix "*" := Int64.mul : Int64Scope.
+Infix "<" := Int64.lt : Int64Scope.
+Notation "x <= y" := (negb (Int64.ltu y x)) (at level 70) : Int64Scope.
+Infix "%" := Int64.mods (at level 70) : Int64Scope.
+Infix "//" := Int64.divs (at level 70) : Int64Scope.
  
 Delimit Scope PtrofsScope with ptrofs.
 Infix "==" := Ptrofs.eq (at level 70) : PtrofsScope.
